WASD & Jump Keys and Camera
Player Basic Controls and Camera Movement

[PlayerInput Manual]
There are three things
* Action Maps
A set of actions that the player can do during the current state of the game, e.g. OnFoot are actions that the player can do when walking around (sliding, run jump etc.)
In summary, a group of actions categorized

* Actions
Where you inputs are handled, keyboard input such as WASD can be used in from the Action Map.
* Properties
Modify properties of each action or ActionMap

[Notes]
- Moved the Ball by X-coordinates to make room for the player
- Created a Prefab folder
- Installed the Input System Package, keep in mind because of this we're disabling Unity default control system due that we know we need to have custom key inputs.

[Player Input Debug]
In Unity, go to Windows > Analysis > InputDebugger
This is used for debugging player input, it will find all devices connected to your computer.

Click on Keyboard to check the WASD keys.

Click on Mouse to check the Camera movement.

[Gravity]
There are gravity tags sticked in the Script folder of both C# files, PlayerMotor.cs and InputManager.cs. If you need to know where gravity is affected with the player, examine these scripts.

[Player Camera]
PlayerLook.cs script is the main camera for the First-Person movement, if you are looking to modify the player camera, it is located in there. The InputManager manages the the PlayerLook.cs so it is related to each other.

~ I don't know if there is an error with the camera because I assigned the <Main Camera> to the PlayerLook script WITHIN the inspector but not in the script. So if you getting errors on the console log about the camera, please tell me.