Currency & RGB - Fundamentals
[Scripts Affected]
Currency User Story has affected the following scripts

* Added
- Edited

* PlayerBalance.cs
- DefaultGun.cs (For Testing)

When shooting the gun, it will increment the money and update the UI, this is just for testing purposes. So everytime you click it just adds money. Only for testing.

RGB User Story has created a script

* GunColour.cs

located in Assets/Scripts/ScriptableObjects
When applied to a gun (or a mesh of a gun),  it will change the Albedo or Main colour of that weapon, for default it is set to constant value of cyan. However gun testing will be implemented once more guns are added in the scene/level.

[What have I edited for other work]

- Edited the crosshair to have some transparency
- Change the enemy colour to red
- Moved some of the script folders including the player-related ones.