Menu Revamp V2
~Canvas Change for Unity~

When selecting the Canvas, the change has occurred to have the scalability of the UI match the resolution of the computer.

This change the following

UI Scale Mode - Scale with Screen Size
Screen Match Mode - Expand (this is because the game will be on PC, change this if we are going mobile or other aspect ratio)

~Class UI~

The 'Soldier' text is just a placeholder, please change this when we added the new classes.

~Upgrade Placeholders~

There are some text that haven't been scripted and are constant static text, things such as the skills or the player/gun stats. They have not be implemented. Accessories also are empty.

~Settings~

The style is outdated but it will be reworked soon.